Adding xmodem.c to github
This is something I put together long ago when there was no way (this was B.K. - Before Kermit) to push files from CP/M up onto a Unix machine other than by what we today would call "paste".  At 300 or 1200 baud, the likelyhood that you'd overrun your buffer was nearly nil, but to that would not work for a binary file.  It is simple code, but it saved an untold number of people from having to re-write it for themselves.